P4PU-507 moved dto
